Update part5d.md
Exercise 5.20 is different in playwright than in cypress... I was about to do extra code to make the user able to edit their blogposts and authors.... In Cypress all it wanted was to see if users can like blogs. I'm going to do the like thing since i believe this is what is wanted.